@@ -6,68 +6,70 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=3.3.0 <4.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  percent_indicator: ^3.0.1
+  percent_indicator: ^4.2.3
   cupertino_icons: ^1.0.6
-  get_it: ^7.2.0
-  provider: ^6.0.1
+  get_it: ^7.6.7
+  provider: ^6.1.2
   connectivity: ^3.0.3
-  dio: ^4.0.0-prev2
+  dio: ^5.6.0
   http: ^0.13.1
-  intl: ^0.17.0
+  #intl: ^0.19.0
   #http: ^1.1.0
-  freezed: ^2.4.1
-  shared_preferences: ^2.2.0
+  freezed: ^2.5.7
+  shared_preferences: ^2.3.1
   fancy_bottom_navigation: ^0.3.3
-  carousel_slider: ^4.0.0-nullsafety.0
-  flutter_staggered_grid_view: ^0.6.2
+  carousel_slider: ^5.0.0
+  flutter_staggered_grid_view: ^0.7.0
   shimmer: ^3.0.0
-  url_launcher: ^6.1.12
-  image_picker: ^1.0.1
-  photo_view: ^0.14.0
-  package_info_plus: ^4.0.2
+  url_launcher: ^6.3.0
+  image_picker: ^1.1.2
+  photo_view: ^0.15.0
+  package_info_plus: ^4.2.0
+  universal_platform: ^1.0.0
+  modal_bottom_sheet: ^2.0.0
 
-  firebase_core: ^2.13.0
-  firebase_messaging: ^14.6.1
-  flutter_local_notifications: ^14.1.0
+  firebase_core: ^3.2.0
+  firebase_messaging: ^15.0.1
+  flutter_local_notifications: ^17.2.2
   navigation_history_observer: ^1.1.0
-  firebase_core_platform_interface: ^4.8.0
+  firebase_core_platform_interface: ^5.2.0
   flutter_svg: any
   flutter_html: any
-  webview_flutter: ^2.0.4
-  phone_number: ^2.0.1
-  country_code_picker: ^2.0.2
+  webview_flutter: ^3.0.1
+  phone_number: ^2.1.0
+  country_code_picker: ^3.0.0
   pin_code_fields: ^8.0.1
-  google_maps_flutter: ^2.3.1
+  google_maps_flutter: ^2.9.0
   flutter_google_places: ^0.3.0
   geocoding: any
-  geolocator: ^9.0.2
-  flutter_typeahead: ^4.6.2
-  flutter_datetime_picker_plus: ^2.0.1
-  dropdown_button2: ^1.4.0
-  flutter_slidable: ^3.0.0
-  floating_bubbles: ^2.2.0
-  path_provider: ^2.0.15
-  permission_handler: ^10.4.3
-  flutter_downloader: ^1.8.3
+  geolocator: ^12.0.0
+  flutter_typeahead: ^5.2.0
+  flutter_datetime_picker_plus: ^2.2.0
+  dropdown_button2: ^2.3.9
+  flutter_slidable: ^3.1.1
+  floating_bubbles: ^2.6.2
+  path_provider: ^2.1.4
+  permission_handler: ^11.3.1
+  flutter_downloader: ^1.11.8
   android_path_provider: ^0.3.0
-  device_info_plus: ^9.0.2
-  share_plus: ^7.0.2
+  device_info_plus: ^10.1.2
+  share_plus: ^10.0.0
   charts_flutter: ^0.12.0
   uuid: any
-  flutter_spinkit: ^5.1.0
-  fluttertoast: ^8.0.9
+  flutter_spinkit: ^5.2.1
+  fluttertoast: ^8.2.8
   flutter_polyline_points: ^1.0.0
   text_divider: ^1.0.0
-  simple_barcode_scanner: ^0.0.8
-  badges: ^3.1.1
-  toggle_switch: ^2.1.0
-  pdf: ^3.8.0
-  printing: ^5.9.0
+  simple_barcode_scanner: ^0.1.1
+  badges: ^3.1.2
+  toggle_switch: ^2.3.0
+  pdf: ^3.11.1
+  printing: ^5.13.1
   open_file: ^3.3.1
 
 dev_dependencies:
